requirements.txt
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
torch>=2.2.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
shap>=0.44.0
plotly>=5.18.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=42.0.2
transformers>=4.37.0
pytest>=8.0.0
httpx>=0.26.0
requests>=2.31.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
audio-recorder-streamlit>=0.0.8
